Fix Markdown link in README
The Markdown link was not rendered correctly in the section about installing the DAP Link app on a Flipper Zero. Adding a newline between the `<summary>` and the expanded text fixes this.

Verified by checking the before and after versions of the formatted README in this PR. GitHub shows a preview of both.